pyproject.toml
[tool.poetry]
name = "lootx"
version = "0.1.0"
description = ""
authors = ["EliseyGodX"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
litestar = { extras = ["standart"], version = "^2.14.0" }
jinja2 = "^3.1.5"
uvicorn = "^0.34.0"
requests = "^2.32.3"
sqlalchemy = "^2.0.37"
python-dotenv = "^1.0.1"
python-ulid = "^3.0.0"
asyncpg = "^0.30.0"
pydantic = "^2.10.6"
redis = "^5.2.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
aiosmtplib = "^3.0.2"
celery = { extras = ["beat"], version = "^5.4.0" }
psycopg2-binary = "^2.9.10"
kapusta = "^0.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
flake8-annotations = "^3.1.1"
flake8-bugbear = "^24.12.12"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
flake8-pyproject = "^1.2.3"
mccabe = "^0.7.0"
pycodestyle = "^2.12.1"
pyflakes = "^3.2.0"
unify = "^0.5"
wemake-python-styleguide = "^1.0.0"
flake8-in-file-ignores = "^0.3.0"


[tool.flake8]
max-line-length = 88
required-plugins = [
    "flake8-annotations",
    "flake8-bugbear",
    "flake8-docstrings",
    "flake8-isort",
    "flake8-pyproject",
    "mccabe",
    "pycodestyle",
    "pyflakes",
    "flake8-in-file-ignores",
]

ignore = [
    "ANN002",
    "B907",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D104",
    "D106",
    "D107",
    "E701",
    "E704",
    "I001",
    "I005",
    "W503",
    "ANN003",
    "ANN101",
    "ANN102",
    "WPS111",
    "WPS214",
    "WPS217",
    "WPS229",
    "WPS235",
    "WPS602",
    "WPS604",
    "WPS420",
    "WPS430",
    "WPS231",
    "WPS232",
    "WPS238",
    "WPS211",
    "WPS336",
]

select = ["ANN", "B", "C", "D", "E", "F", "I", "T4", "W"]
extend-select = ["B9"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
